analyzer: Add support of placement new and improved operator new [PR105948]
Fixed spurious possibly-NULL warning always tagging along throwing
operator new despite it never returning NULL.
Now, operator new is correctly recognized as possibly returning
NULL if and only if it is non-throwing or exceptions have been disabled.
Different standard signatures of operator new are now properly recognized.

Added support of placement new, so that is now properly recognized,
and a 'heap_allocated' region is no longer created for it.
Placement new size is also checked and a 'Wanalyzer-allocation-size'
is emitted when relevant, as well as always a 'Wanalyzer-out-of-bounds'.

gcc/analyzer/ChangeLog:

	PR analyzer/105948
	* analyzer.h (is_placement_new_p): New declaration.
	* call-details.cc
	(call_details::maybe_get_arg_region): New function.
	Returns the region of the argument at given index if possible.
	* call-details.h: Declaration of above function.
	* kf-lang-cp.cc (is_placement_new_p): Returns true if the
	gcall is recognized as a placement new.
	* region-model.cc (region_model::eval_condition):
	Now recursively call itself if one the operand is wrapped in a cast.
	* sm-malloc.cc (malloc_state_machine::on_stmt): Added
	recognition of placement new.

gcc/testsuite/ChangeLog:

	PR analyzer/105948
	* g++.dg/analyzer/out-of-bounds-placement-new.C: New test.
	* g++.dg/analyzer/placement-new.C: Added tests.
	* g++.dg/analyzer/new-2.C: New test.
	* g++.dg/analyzer/noexcept-new.C: New test.
	* g++.dg/analyzer/placement-new-size.C: New test.

Signed-off-by: benjamin priour <priour.be@gmail.com>