analyzer: New option fanalyzer-trim-diagnostics [PR110543]
-fanalyzer-trim-diagnostics=[never|system|<maxdepth>] defaulting to system

This option reduce the noise of the analyzer emitted diagnostics
when dealing with system headers, or simply diagnostic paths that are
too long. The new option only affects the display of the diagnostics,
but doesn't hinder the actual analysis.

Some of the new test cases fail for c++14 though, as the diagnostic paths
are not the same size between c++14 and other c++ versions.

In c++14, get events 1-3 at depth 0 when testing
g++.dg/analyzer/fanalyzer-trim-diagnostics-never.C

Whereas c++17 and c++20 give only 2 events.
Thus c++14 fails when tested.