analyzer: Add tests for future support of ownership tracking
Added tests cases for future options -Wanalyzer-ownership,
-Wanalyzer-partial-ownership-transfer and -Wanalyzer-use-after-move.

The test cases cover use-after-move, acquisition of ownership,
and partial transfer of ownership, such as a user-defined move of only
a subset of the resource owned, or a shallow copy of the owned
resources.

Notably, there still lack test for ownership release, as well as proper
testing of successful ownership's transfer.

Test cases of proper behavior of std::shared_ptr should be added in the
future.

gcc/testsuite/ChangeLog:

	* g++.dg/analyzer/ownership-acquire.C: New test.
	* g++.dg/analyzer/ownership-unique_ptr.C: New test.
	* g++.dg/analyzer/partial-transfer.C: New test.
	* g++.dg/analyzer/shallow-copy.C: New test.
	* g++.dg/analyzer/use-after-move.C: New test.